backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
python-jose==3.5.0
passlib[argon2]==1.7.4
SQLAlchemy==2.0.43
pydantic==2.11.9